\name{nicStructure-package}
\alias{nicStructure-package}
\alias{nicStructure}
\docType{package}
\title{
\packageTitle{nicStructure}
}
\description{
\packageDescription{nicStructure}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{nicStructure}
\packageIndices{nicStructure}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{nicStructure}

Maintainer: \packageMaintainer{nicStructure}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
